🎨 Changed project structure, because all the services in the same package is a bad idea
In each service, we have authService interface for example, and we have
to change the name of the interface in each service.

In the long run, this could be really bad.

So I've changed the structure and put each service and its repositories
in the same package to prevent this.